refactor: rename encrypted certs file to certs.ts and update import paths
