+++ tools/encrypt-certs.ts
@@ -21,8 +21,8 @@ const signerKey = readFileSync(join(__dirname, 'pass.run.dfyb.key'))
 const signerKeyEncrypted = encrypt(signerKey, key)
 output += `export const signerKey = '${signerKeyEncrypted.data}'\n`
 
-writeFileSync(join(__dirname, '..', 'api', 'certs.enc.ts'), output)
+writeFileSync(join(__dirname, '..', 'api', 'certs.ts'), output)
 
-console.log('Successfully encrypted to /pass-models/certs.enc.ts')
+console.log('Successfully encrypted to /pass-models/certs.ts')
 console.log('The following encryption key will only be shown once - make sure to copy it')
 console.log(`Key: ${key.toString('base64')}`)
